move ttest back to stats
